Move accessions search to SeedBankService
